pyproject.toml
[project]
name = "tolvera_poc"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "pydantic (>=2.10.6,<3.0.0)",
    "llms-txt (>=0.0.4,<0.0.5)",
    "pysymbol-llm (>=0.0.5,<0.0.6)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "typer (>=0.15.2,<0.16.0)",
    "openai (>=1.71.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "tolvera (>=0.1.0rc15.post1,<0.2.0)"
]

[tool.poetry.scripts]
tolvera_poc = "tolvera_poc.main:app"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
